backend/requirements-core.txt
# Core dependencies for multiagent system
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Azure core
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-core==1.29.6

# HTTP clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0

# OpenAI
openai==1.5.0

# Security
cryptography==41.0.8
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0

# Utilities
python-json-logger==2.0.7
websockets==12.0
sse-starlette==1.8.2
jinja2==3.1.2
